Normalize empty lists to undefined and use data classes
This simplifies the model as we no longer represents both empty
lists and the non-presence of fields. They are now both deserialized
to the empty list, and never serialized.

Using data classes adds the ability to modify the data while copying it,
the ability to perform structural equility checks, and generates a
usable toString() method.